v2/requirements.txt
# TradeOS V2 — Python Dependencies
# Core
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data
yfinance>=0.2.30
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# HTTP / Scraping
httpx[http2]>=0.24.0
aiohttp>=3.9.0

# ML Models
scikit-learn>=1.3.0
xgboost>=1.7.0
hmmlearn>=0.3.0
arch>=6.0.0

# AI
groq>=0.4.0

# Sentiment / NLP
# trafilatura>=1.6.0  # Optional: for news content extraction
# ddgs>=5.0.0         # Optional: for DuckDuckGo search

# Time zones
pytz>=2023.3

# Misc
websockets>=11.0
certifi>=2024.0.0

# Production server
gunicorn>=21.0.0
